chore: replace useBatch with useBatchV2
