--- demo3PNG/src/hooks/useBatchV2.ts
@@ -1,73 +1,0 @@
-import { useReducer } from "react";
-import { Attributes } from "../types";
-import { shuffle } from "../utils";
-
-type State = Array<Attributes>;
-type Action =
-  | {
-      type: "update";
-      payload: {
-        index: number;
-        entity: Attributes;
-      };
-    }
-  | {
-      type: "shuffle";
-      payload: {
-        total: number;
-      };
-    };
-
-const DEFAULT_TOTAL = 1000;
-
-const generateState = (total: number = DEFAULT_TOTAL): State => {
-  // TODO: shuffle based on supply
-  return new Array(total).fill(0).map(() => shuffle());
-};
-
-const reducer = (state: State, action: Action): State => {
-  switch (action.type) {
-    case "update":
-      return state.map((entity, index) => {
-        if (index === action.payload.index) {
-          return {
-            ...entity,
-            ...action.payload.entity,
-          };
-        }
-        return entity;
-      });
-    case "shuffle":
-      return generateState(action.payload.total);
-    default:
-      return state;
-  }
-};
-
-const useBatchV2 = (total: number = DEFAULT_TOTAL) => {
-  const initialState = generateState(total);
-  const [entities, dispatch] = useReducer(reducer, initialState);
-
-  return {
-    entities,
-    updateEntity: (index: number, entity: Attributes) => {
-      dispatch({
-        type: "update",
-        payload: {
-          index,
-          entity,
-        },
-      });
-    },
-    shuffleEntities: (total: number = DEFAULT_TOTAL) => {
-      dispatch({
-        type: "shuffle",
-        payload: {
-          total,
-        },
-      });
-    },
-  };
-};
-
-export default useBatchV2;
